+++ edipack2triqs/hamiltonian.py
@@ -5,12 +5,13 @@ Hamiltonian and its parameters
 from itertools import product
 from dataclasses import dataclass
 from types import NoneType
-from typing import Union
+from typing import Union, Optional
 
 import numpy as np
 
 import triqs.operators as op
 
+from . import EDMode
 from .util import (is_spin_diagonal,
                    is_spin_degenerate,
                    IndicesType,
@@ -24,8 +25,8 @@ from .bath import BathNormal, BathHybrid, BathGeneral
 class HamiltonianParams:
     """Parameters of the Hamiltonian"""
 
-    # EDIpack exact diagonalization mode (normal, superc, nonsu2)
-    ed_mode: str
+    # Exact diagonalization mode
+    ed_mode: EDMode
     # Non-interacting part of the impurity Hamiltonian
     Hloc: np.ndarray
     # Bath object (None if no bath is present)
@@ -63,7 +64,7 @@ def _is_density_density(U: np.ndarray):
     return True
 
 
-def _make_bath(ed_mode: str,
+def _make_bath(ed_mode: EDMode,
                nspin: int,
                Hloc: np.ndarray,
                h: np.ndarray,
@@ -98,7 +99,8 @@ def parse_hamiltonian(hamiltonian: op.Operator,  # noqa: C901
                       fops_imp_up: list[IndicesType],
                       fops_imp_dn: list[IndicesType],
                       fops_bath_up: list[IndicesType],
-                      fops_bath_dn: list[IndicesType]) -> HamiltonianParams:
+                      fops_bath_dn: list[IndicesType],
+                      f_ed_mode: Optional[EDMode] = None) -> HamiltonianParams:
     """
     Parse a given Hamiltonian and extract parameters from it.
     """
@@ -235,9 +237,9 @@ def parse_hamiltonian(hamiltonian: op.Operator,  # noqa: C901
         assert is_spin_degenerate(h)
         assert is_spin_degenerate(V)
         if (Delta == 0).all():
-            ed_mode = "normal"
+            ed_mode = EDMode.NORMAL
         else:
-            ed_mode = "superc"
+            ed_mode = EDMode.SUPERC
     else:  # nspin == 2
         if not (Delta == 0).all():
             raise RuntimeError(
@@ -246,9 +248,19 @@ def parse_hamiltonian(hamiltonian: op.Operator,  # noqa: C901
             )
         if is_spin_diagonal(Hloc) and \
            is_spin_diagonal(h) and is_spin_diagonal(V):
-            ed_mode = "normal"
+            ed_mode = EDMode.NORMAL
         else:
-            ed_mode = "nonsu2"
+            ed_mode = EDMode.NONSU2
+
+    # Check if the forced ED mode is compatible with the deduced one
+    if (f_ed_mode is not None) and (ed_mode != f_ed_mode):
+        if ed_mode == EDMode.NORMAL:
+            ed_mode = f_ed_mode
+        else:
+            raise RuntimeError(
+                f"Requested exact diagonalization mode {f_ed_mode} "
+                f"is incompatible with the Hamiltonian (must be {ed_mode})"
+            )
 
     bath = _make_bath(ed_mode, nspin, Hloc, h, V, Delta) \
         if nbath_total > 0 else None